internal/acctest: custom framework disappears state function (#36410)
This change adds a new helper function to the `acctest` package,
`CheckFrameworkResourceDisappearsWithStateFunc`, that allows for a custom
function to be provided which constructs the plugin framework state from
the Terraform instance state. This is helpful for resoures where the
default behavior of the `CheckFrameworkResourceDisappears` helper (which
only copies root level string arguments into the plugin framework state)
results in missing inputs for the delete operation. Cases where nested
or non-string arguments must be copied can now rely on custom state functions
local to the service testing package.