tools: aws sdk migration semgrep rules
These rules are intended to assist with AWS SDK for Go migrations by auto-fixing certain patterns. The resulting code will be incomplete and may not compile. This tool will not automate migrations, but can alleviate some of the repetitive tasks required to updgrade a service.

The command below executes a dry run. Replace the `<service>` placeholder with the service to be migrated.

```console
% semgrep --config tools/.semgrep-awssdk-migration.yml --dryrun internal/service/<service>
```

Once comfortable with the proposed changes, pass the `--autofix` flag to apply automated fixes.

```console
% semgrep --config tools/.semgrep-awssdk-migration.yml --autofix internal/service/<service>
```