chore: add awssdkpatch make targets
awssdkpatch - Install awssdkpatch
awssdkpatch-gen - Generate a patch file using awssdkpatch
awssdkpatch-apply - Apply a patch generated with awssdkpatch

The first command is called without parameters. The `-gen` and `-apply` targets require PKG to be set in order to determine which service the patch should be generated for and/or applied to.