Improve import docs for aws_bcmdataexports_export resource
This resource was added in v5.47.0 and I took it for a test run. We have
an existing v2.0 export that was manually created, and I wanted to
import it into terraform management.

It wasn't clear from the docs what format the ID should be in, but by
reading the provider source I could see it was the export ARN. I thought
I'd mention that in the docs to help others.

Interestingly, finding the export ARN from the AWS web console wasn't
easy. The URL was the only option I could find.