Fix for default_tags and tags on EC2 block devices
`default_tags` support for EC2 root block devices was introduced in
hasicorp/terraform-provider-aws#33769, which is great ... except it
makes the mix of `default_tags` and `tags` on a `root_block_device`
perpetually showing drift even when there is none.

The solution is to not remove the default tags config from the volume
tags in ec2_instance.

This diff fixes hashicorp/terraform-provider-aws#36448 for us in our
environment (which references hashicorp/terraform-provider-aws#33769 as
the change that introduced this bug).

This diff possibly fixes hashicorp/terraform-provider-aws#36706 too, but
we weren't hitting this issue.