awssdkpatch: enum type validation
This enhancement adds patches to migrate validation functions for enum types to the `enum.Validate` function.

Due to a `gopatch` limitation, the field which the validation function is assigned to (`ValidateFunc`) cannot be migrated to the proper field (`ValidateDiagFunc`). The generated patches include a comment indicating that this step will need to be done manually once the patch is applied, at least until the underlying `gopatch` limitation is addressed.

Related section in the AWS SDK migration guide:
- https://hashicorp.github.io/terraform-provider-aws/aws-go-sdk-migrations/#validation-functions