Adjust documentation of enable_http_endpoint option of RDS cluster
The documentation for this option has been long untouched and some things have changed.
It is possible now to use Data API for provisioned Aurora clusters, however this is available
only for certain versions and in certain regions.

This also adds a note of options that are not compatible with enable_http_endpoint in Terraform.