r/aws_bedrockagent_agent: use configured `prepare_agent` value
This changes removes logic in the read operation which adjusted the `prepare_agent` argument based on whether or not the agent status was `Prepared`. Now the configured `prepare_agent` value is always used, regardless of the status of the agent. If omitted, the default value of `true` is inherited, meaning that by default agents will always be prepared as part of the create or update operation. The `prepare_agent` value is now also set to `true` when an agent is imported.

```console
% make testacc PKG=bedrockagent TESTS=TestAccBedrockAgentAgent_
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go1.22.2 test ./internal/service/bedrockagent/... -v -count 1 -parallel 20 -run='TestAccBedrockAgentAgent_'  -timeout 360m

--- PASS: TestAccBedrockAgentAgent_basic (19.66s)
--- PASS: TestAccBedrockAgentAgent_full (19.67s)
--- PASS: TestAccBedrockAgentAgent_tags (35.51s)
--- PASS: TestAccBedrockAgentAgent_update (37.46s)
PASS
ok      github.com/hashicorp/terraform-provider-aws/internal/service/bedrockagent       42.493s
```