d/aws_route53_zone: permit both `name` and `zone_id` arguments
As part of the migration of the `route53` service to AWS SDK V2, an `ExactlyOneOf` constraint was added to the `name` and `zone_id` arguments in the `aws_route53_zone` data source. This is technically a regression as the data source functioned correctly in earlier versions when `name` is set and `zone_id` is an empty string. This change reverts the constraint and adds an acceptance test to cover this potential configuration.

```console
% make testacc PKG=route53 TESTS=TestAccRoute53ZoneDataSource_
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go1.22.2 test ./internal/service/route53/... -v -count 1 -parallel 20 -run='TestAccRoute53ZoneDataSource_'  -timeout 360m

--- PASS: TestAccRoute53ZoneDataSource_id (91.23s)
--- PASS: TestAccRoute53ZoneDataSource_name_idEmptyString (95.31s)
--- PASS: TestAccRoute53ZoneDataSource_name (101.49s)
--- PASS: TestAccRoute53ZoneDataSource_serviceDiscovery (102.80s)
--- PASS: TestAccRoute53ZoneDataSource_vpc (126.11s)
--- PASS: TestAccRoute53ZoneDataSource_tags (127.60s)
PASS
ok      github.com/hashicorp/terraform-provider-aws/internal/service/route53    132.691s
```