Allow `package_type` to be empty when `image_uri` is set in Lambda functions
This commit introduces changes that allow the `package_type` parameter in AWS Lambda functions to be empty when `image_uri` is set. Accordingly, the parameter now defaults to `Image` if `image_uri` is set. The necessary tests have been added to verify this behaviour and the documentation has been updated to reflect this change. Modifications have been made internally to validate the `package_type` and `image_uri` combination.