Clarify the instance type to use for Aurora Serverless v2
When creating an Aurora Serverless v2 cluster in Terraform, users
must add an instance to the cluster. The documentation points to the
AWS documentation page on the currently available instance classes,
but this documentation does not include `db.serverless`.

The Aurora Serverless v2 documentation itself also has scant mention
of it.

Including the information here as well ought to help users.