r/aws_chime_voice_connector_group: handle NotFound error on delete
The delete operation calls an update when the `connector` argument is configured, which was not handling cases where the group did not exist (deleted out of band). This change prevents this error from propagating back to the user, and instead just removes the resource from state when the group cannot be found during the delete operation.