fix(opensearchserverless): security_policy policy attribute doesn't normalize json strings
The `policy` attribute is a string type, but requires the string to be
valid json. The user can input json with the keys in any order, but the
value stored in state will have the keys in order causing an error
message on apply (see linked issue for details).

This resource is very similar to the `access_policy` resource which uses
`jsontypes.Normalize` as the type instead of `types.String`. I have
copied most of the implementation from `access_policy` and added tests
which test using a string value instead of `jsonencode`.

fixes #38603