r/aws_cloudformation_stack_set_instance: fix id constructor
The create method incorrectly assumed that when `deployment_targets` are provided, organizational unit IDs must be provided in that block, and attempted to use those values in the multi-part key stored for the `id`. This is not the case for self-managed stack sets, which allow `deployment_targets` to be configured with an empty list of organzation unit IDs. The provider will now fall back to `account_id` for this section of the identifier instead.