ci(semgrep): add avoid-errs-Must check
This check will match on any use of `errs.Must(...)` within `internal/service/**/`. Errors should be handled explicitly in service packages to avoid the possibility of crashing the provider during execution. This function was originally intended for use while initializing service clients during the `ConfigureProvider` operation, where a panic is the appropriate course of action.