.ci/tools: `go get -u github.com/katbyte/andreyvit-diff@v0.0.1`
At some point this indirect dependency of `github.com/katbyte/terrafmt` was updated to point to a commit beyond the latest published tag on the source repository (`v0.0.1`). This commit explicitly fetches the depedency at the `v0.0.1` tag in an attempt to resolve installation issues for contributors who cannot use the default `GOPROXY` settings which handle resolving this "yet-to-be-published" pseudo-version. The latest tagged version of `terrafmt` also [uses `v0.0.1` of this library](https://github.com/katbyte/terrafmt/blob/v0.5.4/go.mod#L12).