Handle redis -> valkey engine updates on replication_group
This was a bit more tricky because AWS will allow modify requests for
the replication group when only the engine attribute is changed but this
actually ends up resulting in no actual upgrade operation.

What I've discovered is that for this process to be successful, the
engine_version attribute needs to be explicitly provided as well.

Signed-off-by: Aurel Canciu <aurel.canciu@nexhealth.com>