docs: remove non-usable workflow type
### Description

The documentation of the resource [`ami_imagebuilder_workflow`](https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/imagebuilder_workflow)  mentions for the argument `type`
>  Valid values: BUILD, TEST, DISTRIBUTION.

Image Builder handles the  workflow type DISTRIBUTION, it cannot be configured by users.
When trying to do so, for example using the code snippet provided as part of #39785 

```hcl
resource aws_imagebuilder_workflow codebuild-projects {
  name = "CodeBuildProjectsDistributionWorkflow"
  version = "1.0.0"
  type = "DISTRIBUTION"
  data = file("${path.module}/codebuild-projects-distribution.workflow.yaml")
}
```

the API returns the error ` The value supplied for parameter 'type' is not valid. Custom distribution workflows are not supported.`

This pull request is to update the documentation for the resource.

### Relations

Closes #39785 

### References

[Here](https://github.com/hashicorp/terraform-provider-aws/issues/39785#issuecomment-2424187935) I explained the current situation.

### Output from Acceptance Testing

Not required. Only documentation changed.