internal/framework/types: surface ARN validation errors (#40008)
Previously the `ARNType`s `ValueFromString` method parsed the string value and set the value to Unknown in cases where validation failed. This caused a plan time failure indicating a difference between the planned value (unknown) and config value (known string). This obfuscated the underlying error which is actually an invalid ARN value.

Invalid ARNs provided in a configuration will now produce errors similar
to the following:

```
│ Error: Invalid ARN Value
│
│   with aws_bedrockagent_agent.test,
│   on main.tf line 16, in resource "aws_bedrockagent_agent" "test":
│   16:   agent_resource_role_arn = "INVALID-ARN"
│
│ The provided value cannot be parsed as an ARN.
│
│ Path: agent_resource_role_arn
│ Value: INVALID-ARN
```