r/aws_lb_listener: remove `tcp_idle_timeout_seconds` default
The `tcp_idle_timeout_seconds` default value of `350` has been removed in favor or marking this argument optional and computed. In practice, this means the default value of `350` is still read and stored in state for protocols which support this feature, but unless explicitly configured by the user, the `ModifyListenerAttributes` API will not be called during create or update operations.

This change fixes a regression introduced by the addition of the `tcp_idle_timeout_seconds` argument with a default value, which caused any existing configuration using the `aws_lb_listener` resource to now require the calling principal to include the `elasticloadbalancing:ModifyLoadBalancerAttributes` IAM action in an attached IAM policy. By only sending this value when explicitly configured, existing configurations can continue to operate as expected with no modifications to the IAM permissions of the calling principal.

To confirm the API is no longer being called when `tcp_idle_timeout_seconds` is not explicitly configured, I've grepped the debug logs of the same configuration applied with `v5.74.0` of the AWS provider and a locally built binary from this branch. The former includes a result with the API request body where the TCP idle timeout is set to the default of `350`. The latter includes no result indicating the `ModifyListenerAttributes` API was not called, as expected.

__`v5.74.0`__:

```console
% rg "Action=ModifyListenerAttributes" apply-old.out
2672:  | Action=ModifyListenerAttributes&Attributes.member.1.Key=tcp.idle_timeout.seconds&Attributes.member.1.Value=350&ListenerArn=arn%3Aaws%3Aelasticloadbalancing%3Aus-west-2%3A<redacted>%3Alistener%2Fnet%2Fjb-test%2Fb0f926497af0c80f%2F819828f77724007f&Version=2015-12-01
```

__This branch__:

```console
% rg "Action=ModifyListenerAttributes" apply.out
```

```console
% make testacc PKG=elbv2 TESTARGS="-run=TestAccELBV2Listener_.*_basic"
make: Verifying source code with gofmt...
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go1.23.2 test ./internal/service/elbv2/... -v -count 1 -parallel 20  -run=TestAccELBV2Listener_.*_basic -timeout 360m
2024/11/06 16:43:21 Initializing Terraform AWS Provider...

--- PASS: TestAccELBV2Listener_Gateway_basic (191.64s)
--- PASS: TestAccELBV2Listener_Network_basic (214.10s)
--- PASS: TestAccELBV2Listener_Application_basic (229.81s)
PASS
ok      github.com/hashicorp/terraform-provider-aws/internal/service/elbv2      236.141s
```