chore: rm changelog entry
The AWS SDK for Go release which contained these new `deployment_type` valid values has been available for some time in the AWS provider. To avoid misleading users about which version in which these values first became available, a changelog entry will be omitted for this change in favor of documentation-only updates.