.github/workflows: address breaking changes to `slack-github-action` (#40213)
The `webhook-type` argument is now required. The `webhook` argument is optional and can also be configured by setting the `SLACK_WEBHOOK_URL` environment variable, but I've migrated to the explicit argument to align with the documentation for the release. See the [release notes](https://github.com/slackapi/slack-github-action/releases/tag/v2.0.0) for complete details on the breaking changes.