ssm: align `with_decryption` default with `data.aws_ssm_parameter`
Terraform does not plan for ephemeral resources, data sources,
or providers. As such, the Plugin Framework does not implement
default handlers for these objects.

To provide a consistent user experience with the `data.aws_ssm_parameter` data
source, we default the `with_decryption` attribute to `true` in its ephemeral implementation.