.github/workflows: fix registry version check comparison
The version returned from the `curl` request to the registry will contain only the version number, while the release tag is prefixed with `v`.

For example, here is what is stored in the `LATEST_VERSION` environment
variable:

```console
% curl -s "https://registry.terraform.io/v2/providers/323/provider-versions/latest" | jq -r '.data.attributes.version'
5.79.0
```

This changes the conditional to apply a prefix to the version number fetched from the registry to allow an accurate comparison of the two values.