r/aws_ecs_service: availability_zone_rebalancing attribute shouldn't have a default
An attributed availability_zone_rebalancing that was introduced in PR https://github.com/hashicorp/terraform-provider-aws/pull/40225
is defined as an optional, but also has a default value. For configurations
where this flag is not defined, plan with `refresh=false` results in
a change that would add this attribute to the state - which is not what
"optional" should mean.

This is to address https://github.com/hashicorp/terraform-provider-aws/issues/40349