autoscaling: safeguard `availability_zone_distribution` and alphabetize schema
In the Create, we're safeguarding `availability_zone_distribution` to set the value only if it really has an entry in the backing array.

In the Read, because `availability_zone_distribution` is of type list in the schema, but flattening returns a map, we wrap it back into a list.

Moved `availability_zone_distribution` up in the schema definition to alphabetize it.