r/aws_kms_custom_key_store: remove `xks_proxy_authentication_credential` flattener
Removes the flattener for `xks_proxy_authentication_credential` given only one of the two nested arguments is returned from the AWS API. This was done as a precaution to eliminate potential failure modes in writing the returned value given we cannot acceptance test these changes. If drift detection on the nested `access_key_id` argument is desired by the community, it will be added in a future release.