r/aws_macie2_invitation_accepter: properly set `invitation_id` during creation (#41163)
This fixes a regression introduced during migration of the `macie2` service to AWS SDK V2. Specifically, the `invitationID` variable inside the create retrier shadows a variable of the same name at the root of the create function. This led to the `InvitationId` value always being sent as an empty string in the `AcceptInvitation` API request, despite being successfully retrieved from the `ListInvitations` API. The variable from the outer scope is now used inside the retrier's scope (rather than a new variable), which should return the original intended behavior.

Note: Acceptance tests for this resource require out of band email verification for the Macie2 service, which maintainers do not currently have the capability to do in our testing accounts. As such, this implementation is best effort and we ask for community help in verifying the fix.