Fix listener attributes not being updated
PR #40736 added support for many new listener attributes, however the
resource update code was only written to call ModifyListenerAttributes if
the tcp_idle_timeout_seconds attribute had been updated. This would
result in the terraform plan saying that it would update the attributes,
but the change would not actually get applied.

This change fixes that by making it so that ModifyListenerAttributes is
called whenever any of the attributes has been changed.

Fixes issue #40986.