docs: group generated tag test options with headings
As tag test generation has expanded across the provider, the number of parameters which can be passed to the tags test annotation has grown significantly. Given the number of options, the contributor documentation on this section has become difficult to read in its entirety. This change bundles the various parameters into related groups and adds headings to break up the available options into smaller standalone sections.