Add RDS Aurora Global Cluster Data Source
Unfortunately, I created this before seeing that #37286 existed. However
it is currently out of sync, so perhaps this is still warranted.

I did start with the latest framework type coming from skaff, but had
some issues with tagging, and thought it would be better to keep
consistent with the matching resource anyway, so the sdk type is used
here.

References:

- #13289
- #20171