r/aws_account_region: skip Enabled/DisableRegion calls when in-progress
This change modifies the behavior when a region is already in the process of becoming enabled or disabled. Previously, attempts to enable an already enabled region or a region in the process of being enabled would cause a hard error.

```
│ Error: enabling Account Region (ap-southeast-3): operation error Account: EnableRegion, https response error StatusCode: 400, RequestID: 04e104c7-cafd-4f22-a659-da79bfa8b836, ValidationException: Unable to switch region status because of its current opt-in status.
│
│   with aws_account_region.test,
│   on main.tf line 13, in resource "aws_account_region" "test":
│   13: resource "aws_account_region" "test" {
```

Now, the create and update operations will account for in-progress operations and proceed to the waiting logic instead of attempting to enable or disable a region again. This allows for "adoption" of existing enabled or disabled resources without an import. While this is a non-standard pattern for Terraform, the fact that all regions are _already_ pre-existing in the account make it a reasonable trade off to avoid unexpected behavior when enabling a region times out and the operation is re-attempted.