fix: fix #27674 - resource/aws_apigatewayv2_stage default throttling/burst limit
The burst limit / rate limiting of API Gateway are **not** zero by default, but are rather not provided
at all, when absent.

Ref: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-apigateway-stage-methodsetting.html
Ref: https://github.com/hashicorp/terraform-provider-aws/issues/27674

This change is based on suggestions by @krzysztofdrys

Tests provided had to be written as unit-tests from scratch, since I do not have access to an AWS account
suitable for running the test suite against.