r/aws_ssm_parameter: remove `overwrite` deprecation
After assessing community feedback on the intended removal, we have determined the impact to existing use cases relying on this behavior would be too significant to move forward with in `v6`. The deprecation message is now removed from the resource schema and registry documentation, and the argument can continue to be used with the pre-existing behavior.