r/aws_sns_topic_subscription: handle eventually consistent subscription deletion (#42093)
Previously the `Read` operation relied solely on the `GetSubscriptionAttributes` API, which is eventually consistent and may not correctly identify scenarios when the topic or subscription is deleted out of band. This change adds a call to the `ListSubscriptionsByTopic` API prior to the existing read logic to ensure both the topic and subscription are currently active. When a `NotFound` error is returned or the subscription is not in the list of returned subscriptions, the resource is removed from state and a re-creation will be triggered on the next `apply`.

```console
% make testacc PKG=sns TESTS=TestAccSNSTopicSubscription_
make: Verifying source code with gofmt...
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go1.23.7 test ./internal/service/sns/... -v -count 1 -parallel 20 -run='TestAccSNSTopicSubscription_'  -timeout 360m -vet=off
2025/04/02 10:45:34 Initializing Terraform AWS Provider...

--- PASS: TestAccSNSTopicSubscription_filterPolicyScope_policyNotSet (3.77s)
--- PASS: TestAccSNSTopicSubscription_email (17.52s)
--- PASS: TestAccSNSTopicSubscription_autoConfirmingEndpoint (42.54s)
--- PASS: TestAccSNSTopicSubscription_disappears_TopicExternal (50.90s)
--- PASS: TestAccSNSTopicSubscription_disappears (51.40s)
--- PASS: TestAccSNSTopicSubscription_disappears_Topic (51.53s)
--- PASS: TestAccSNSTopicSubscription_basic (51.89s)
--- PASS: TestAccSNSTopicSubscription_deliveryPolicy (72.32s)
--- PASS: TestAccSNSTopicSubscription_rawMessageDelivery (72.41s)
--- PASS: TestAccSNSTopicSubscription_filterPolicy (73.24s)
--- PASS: TestAccSNSTopicSubscription_firehose (75.93s)
--- PASS: TestAccSNSTopicSubscription_autoConfirmingSecuredEndpoint (85.98s)
--- PASS: TestAccSNSTopicSubscription_redrivePolicy (106.94s)
--- PASS: TestAccSNSTopicSubscription_filterPolicyScope (154.09s)
PASS
ok      github.com/hashicorp/terraform-provider-aws/internal/service/sns        160.822s
```