internal/framework/flex: skip planned unknown values in `Diff`
Previously a computed or optional/computed attribute in state without a `UseStateForUnknown` plan modifier would cause `flex.Diff` to indicate the plan has changes due to the previously known value now being known after apply. In this scenario we now skip this field as it can't be ensured whether or not the value has changed.