provider: deprecate support for s3 global endpoints (#42375)
In `v6.0.0` support for the `s3_us_east_1_regional_endpoint` provider argument, and the ability to use the global S3 endpoint is being deprecated. This argument will be removed in `v7.0.0`.

If `s3_us_east_1_regional_endpoint` is configured with a value of `legacy`, a warning diagnostic will now be returned indicating that support for the global S3 endpoint is deprecated and instructions on how to validate that the regional S3 endpoint is functional in a given AWS environment.