internal/sdkv2/types/nullable: refine `TypeNullableBool` validation (#42434)
Arguments of the type `TypeNullableBool` will now only accept one of `""` (empty string), `true`, or `false`. An upgrade guide entry is included enumerating the resources and arguments which are affected by this validation change.

As the `ValueBool()` function has not changed (still using `strconv.ParseBool()`), any `"1"`s or `"0"`s in state will still work as `true` or `false`, respectively.