internal/vcr: add patch file to support enabling VCR
This patch file can be used to automate the addition of VCR support
across services. As additional wrapper methods are added (such as
wrapping state change data structures) additional patches will be
included in this file.

To apply the patch to an entire service:

```console
% gopatch -p internal/vcr/vcr.patch ./internal/service/<name>/...
```

Alternatively, the patch can be applied to a single file:

```console
% gopatch -p internal/vcr/vcr.patch ./internal/service/rds/cluster_test.go
```

Pass the `-d` flag to display the diff to stdout without applying any
changes:

```console
% gopatch -d -p internal/vcr/vcr.patch ./internal/service/rds/cluster_test.go
```