internal/vcr: add vcr support semgrep rules (#42855)
These rules will apply autofixes that enable VCR support for a given service package. At this time the intent is to manually apply the rules one service at a time and run acceptance tests (both with and without VCR enabled) to verify no breakage has occurred.

A new Make target, `make semgrep-vcr`, has been added to simplify how this process is completed.

```console
% PKG=logs make semgrep-vcr
```
Because certain autofixes apply changes inside blocks which are modified
by other rules, this command may need to be run twice to fully complete
the process. Additionally, the resulting code may need to be reformatted
and imports may need to be tidied, depending on the scope of the change.

```console
% make fmt
```

```console
% goimports -w internal/service/logs/
```