internal/generate/tagstests: enable go-vcr support
This change makes the following VCR-aware replacements in the tagstest templates for resources and data sources:

- `resource.Test` to `acctest.Test`
- `resource.ParallelTest` to `acctest.ParallelTest`
- `sdkacctest.RandomWithPrefix` with `acctest.RandomWithPrefix`

In order to be fully support go-vcr, the "check exists" and "check
destroy" helpers also need to include a `testing.T` argument. The
generator already includes optional `existsTakesT` and `destroyTakesT`
annotations, which can be set on new resources as they become VCR
enabled over time.