r/aws_s3_bucket_logging(test): fix flaky test check
The `target_grant.grantee.display_name` attribute is not always consistently populated when a grant is created via a canonical user ID. This was manifesting in acceptance tests as either a failed check when assuming the expected value of the `display_name` attribute, or failed import verification when a previously known value returned as empty on a subsequent read operation. To avoid these inconsistent failure modes, direct checks on this computed attribute have been removed, and the attribute has been ignored during import verification.