Add compaction options to to r/aws_glue_catalog_table_optimizer
Adds `compaction_configuration` as a configurable block for Glue Table
optimizers.

This change also makes it mandatory for optimizers with `type =
compaction`, which may break some existing configurations. This also
adds validation rules so that optimizer configuration blocks must match
the type of optimizer.

The reason for making `compaction_configuration` mandatory is that the
Glue API now always returns a `CompactionConfiguration` block with
strategy set to `binpack` if no options are passed upon creation.
If `compaction_configuration` is optional and missing, this results in
an entire block that is absent from Terraform configuration but gets
returned in state. This cannot be overcome with a PlanModifier because
the diff affects the block count (0 vs. 1)

The most reliable way to prevent the Terraform config from conflicting
with the API response seems to be mandating that the fields always
present in the response are always present in the config.