internal/generate/identitytests: allow optional attributes to be set in tests
This change introduces a new argument to the `@IdentityAttribute` annotation, `testNotNull`. This can be set to `true` when an optional identity attribute is expected to be non-null in the generated test configuration. This is necessary to support testing of resources with multiple optional identity attributes where at least one is required.