fix[ec2_instance]: Handle invalid utf8 in Read
When reading the UserData for an instance, if the decoded user data isn't valid UTF-8, then store
the base64 value in user_data_base64 instead of the raw value in user_data.

This ensures that we don't store an invalid string in the `user_data` field, resulting in invalid
json.

In particular, this allows importing an aws_instance that has binary UserData without resulting in
an invalid state.

Fixes: #44506