internal/generate/tagrislookup: add website docs
This adds a new guide titled `Tag Policy Enforcement` which includes details about setting up tag policies in the management organization account, enabling enforcement in the AWS provider, and a cross reference mapping tag resource types to Terraform resource types. The cross reference is triggered from the CSV in the internal/tags/tagris directory which is also the source of truth for the lookup used in the required tag validation interceptor.